requirements.txt
discord.py
dotenv-python
